Proof of concept: launch Dolphin from inside Randovania
This makes the following changes:
- Dolphin connector can now have a dolphin command configured, which it
will prompt for on creation
- If a command is configured, "start dolphin" and "stop dolphin" menu
entries attached to the connector allow a dolphin subprocess to be
managed

This fixes issues with ptrace permissions on linux by ensuring that the
dolphin process is a child of RDV and thus traceable. The custom command
allows for launching dolphin-compatible emulators like primehack or
retroarch.

No attempt is made to automatically figure out the dolphin commname
(which DME uses to find it); if it's something other than "dolphin-emu",
the user needs to set the DME_DOLPHIN_PROCESS_NAME envar.